pyproject.toml
[project]
name = "auth0_Flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "flask>=3.1.3",
    "pyjwt[crypto]>=2.11.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
]
